backend/requirements.txt
fastapi 
uvicorn 
sqlalchemy 
psycopg2-binary 
pgvector 
pydantic-settings 
httpx 
python-dotenv 
python-multipart
